Improve card formatting step
Prior to this commit, this was done solely on pre-save hook. Now it's done when
a sync card *to* trello action occurs as well.

This also fixes the card description region computation. Prior to this commit,
this was computing without narrowing the card region first (which in some edge
case resulting in miscomputed region).

Related #407